requirements.txt
discord.py
gspread
google-auth
pytz